Filter click events to toggle the pane
To close the pane when the user clicks outside, the `focusout ` event
was being used. This resulted in the click event on the Provide Feedback
link to be ignored. Removed the focusout approach and added logic to
listen to all clicks on the document / nested iframes to close the pane.

Fixes #59